+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+## [0.3.0] - 2020-11-26
+
+- Fixed a bug that the End Date could not be the same as de Initial Date.
+- Added the possibility to open the calendar with dates already selected (One date, two dates, or none).
+
 ## [0.2.0] - 2020-11-25
 
-- Added the possibility to show disabled dates the First and Last Month, these dates are the dates that are previous `minDate` and post `maxDate`
+- Added the possibility to show disabled dates the First and Last Month, these dates are the dates that are previous `minDate` and post `maxDate`.
 - Fix a bug that the next Month displayed with empty values when the `maxDate` occurs on the last week of the month.
 
 ## [0.1.1] - 2020-10-28
@@ -12,5 +17,5 @@
 
 - Added calendar full page with a scroll in the vertical.
 - Added a select a range date with special colors.
-- Added capture of the date selected
-- Change calendar with Locale
+- Added capture of the date selected.
+- Change calendar with Locale.